fix minor issues in campaign
