+++ src/hooks/wallets/use-connected-wallet.ts
@@ -181,7 +181,7 @@ export const useConnectedXrplWallet = () => {
                     })
                     .catch(() => {
                       close();
-                      resolve(e.data);
+                      reject(e.data);
                     });
                 }
               )
