+++ src/locales/en/translation.json
@@ -254,8 +254,8 @@
   "total-bridge": "Total",
   "Bridge": "Bridge",
   "tooltip-text-enough-xrp": "If you already have enough $XRP in your Root Network wallet, go to the next step!",
-  "Bridge confirmed!": "Bridge confirmed!",
-  "bridge-success-message": "Successfully bridged your $XRP to The Root Network",
+  "Bridge confirmed!": "Bridge request success!",
+  "bridge-success-message": "Successfully requested the bridge. Usual bridge transaction takes 15 minutes. Please wail until you receive your XRP to the Metamask or Futurepass.",
   "Bridge failed!": "Bridge failed",
   "bridge-fail-message": "Bridge your $XRP has failed. Please try again.",
   "Continue to add liquidity": "Continue to add liquidity",
