+++ src/locales/ko/translation.json
@@ -267,8 +267,8 @@
   "total-bridge": "총액",
   "Bridge": "브릿지하기",
   "tooltip-text-enough-xrp": "이미 The Root Network 지갑에 $XRP가 충분히 있다면, 다음 단계로 넘어가세요!",
-  "Bridge confirmed!": "브릿지를 완료했습니다.",
-  "bridge-success-message": "$XRP를 The Root Network로 성공적으로 브릿지했습니다.",
+  "Bridge confirmed!": "브릿지를 요청했습니다.",
+  "bridge-success-message": "브릿지 요청에 성공했습니다. 일반적인 XRP 브릿지는 15분 정도 소요됩니다. 루트 네트워크의 메타마스크나 Futurepass에서 자산을 확인 후 다음단계를 진행해 주세요.",
   "Bridge failed!": "브릿지에 실패했습니다.",
   "bridge-fail-message": "$XRP를 The Root Network로 브릿지하는데 실패했습니다. 다시 시도해주세요.",
   "Continue to add liquidity": "유동성 공급 이어서하기",
